+++ packages/iTwinUI-react/src/core/ComboBox/ComboBoxMenuItem.tsx
@@ -1,3 +1,7 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
 import cx from 'classnames';
 import React from 'react';
 import { MenuItemProps } from '../Menu';
@@ -25,11 +29,12 @@ export const ComboBoxMenuItem = React.memo(
         ...rest
       } = props;
 
-      const { focusedIndex } = useSafeContext(ComboBoxStateContext);
+      const { focusedIndex, enableVirtualization } = useSafeContext(
+        ComboBoxStateContext,
+      );
 
       const focusRef = (el: HTMLLIElement | null) => {
-        // will need to check for virtualization here too
-        if (focusedIndex === index) {
+        if (!enableVirtualization && focusedIndex === index) {
           el?.scrollIntoView({ block: 'nearest' });
         }
       };