+++ packages/iTwinUI-react/src/core/Table/Table.tsx
@@ -28,7 +28,7 @@ import { useTheme, CommonProps, useResizeObserver } from '../utils';
 import '@itwin/itwinui-css/css/table.css';
 import SvgSortDown from '@itwin/itwinui-icons-react/cjs/icons/SortDown';
 import SvgSortUp from '@itwin/itwinui-icons-react/cjs/icons/SortUp';
-import { getCellStyle } from './utils';
+import { getCellStyle, getStickyStyle } from './utils';
 import { TableRowMemoized } from './TableRowMemoized';
 import { FilterToggle, TableFilterValue } from './filters';
 import { customFilterFunctions } from './filters/customFilterFunctions';
@@ -39,6 +39,7 @@ import {
   useSubRowSelection,
   useResizeColumns,
   useColumnDragAndDrop,
+  useStickyColumns,
 } from './hooks';
 import {
   onExpandHandler,
@@ -479,6 +480,7 @@ export const Table = <
     useSelectionCell(isSelectable, selectionMode, isRowDisabled),
     useColumnOrder,
     useColumnDragAndDrop(enableColumnReordering),
+    useStickyColumns,
   );
 
   const {
@@ -495,6 +497,7 @@ export const Table = <
     gotoPage,
     setPageSize,
     flatHeaders,
+    visibleColumns,
   } = instance;
 
   const ariaDataAttributes = Object.entries(rest).reduce(
@@ -507,7 +510,15 @@ export const Table = <
     {} as Record<string, string>,
   );
 
-  const areFiltersSet = allColumns.some((column) => column.filterValue != null && column.filterValue !== '');
+  const areFiltersSet = allColumns.some(
+    (column) => column.filterValue != null && column.filterValue !== '',
+  );
+
+  const showFilterButton = (column: HeaderGroup<T>) =>
+    (data.length !== 0 || areFiltersSet) && column.canFilter;
+
+  const showSortButton = (column: HeaderGroup<T>) =>
+    data.length !== 0 && column.canSort;
 
   const onRowClickHandler = React.useCallback(
     (event: React.MouseEvent, row: Row<T>) => {
@@ -656,6 +667,34 @@ export const Table = <
     [getPreparedRow],
   );
 
+  const updateStickyState = () => {
+    if (!bodyRef.current || flatHeaders.every((header) => !header.sticky)) {
+      return;
+    }
+
+    if (bodyRef.current.scrollLeft !== 0) {
+      dispatch({ type: TableActions.setScrolledRight, value: true });
+    } else {
+      dispatch({ type: TableActions.setScrolledRight, value: false });
+    }
+
+    // If scrolled a bit to the left looking from the right side
+    if (
+      bodyRef.current.scrollLeft !==
+      bodyRef.current.scrollWidth - bodyRef.current.clientWidth
+    ) {
+      dispatch({ type: TableActions.setScrolledLeft, value: true });
+    } else {
+      dispatch({ type: TableActions.setScrolledLeft, value: false });
+    }
+  };
+
+  React.useEffect(() => {
+    updateStickyState();
+    // Call only on init
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
   return (
     <>
       <div
@@ -672,82 +711,108 @@ export const Table = <
             { [`iui-${density}`]: density !== 'default' },
             className,
           ),
-          style,
+          style: {
+            minWidth: 0, // Overrides the min-width set by the react-table but when we support horizontal scroll it is not needed
+            ...style,
+          },
         })}
         {...ariaDataAttributes}
       >
-        <div className='iui-table-header' ref={headerRef}>
-          {headerGroups.slice(1).map((headerGroup: HeaderGroup<T>) => {
-            const headerGroupProps = headerGroup.getHeaderGroupProps({
-              className: 'iui-row',
-            });
-            return (
-              <div {...headerGroupProps} key={headerGroupProps.key}>
-                {headerGroup.headers.map((column, index) => {
-                  const columnProps = column.getHeaderProps({
-                    ...column.getSortByToggleProps(),
-                    className: cx(
-                      'iui-cell',
-                      { 'iui-actionable': column.canSort },
-                      { 'iui-sorted': column.isSorted },
-                      column.columnClassName,
-                    ),
-                    style: { ...getCellStyle(column, !!state.isTableResizing) },
-                  });
-                  return (
-                    <div
-                      {...columnProps}
-                      {...column.getDragAndDropProps()}
-                      key={columnProps.key}
-                      title={undefined}
-                      ref={(el) => {
-                        if (el && isResizable) {
-                          columnRefs.current[column.id] = el;
-                          column.resizeWidth = el.getBoundingClientRect().width;
-                        }
-                      }}
-                    >
-                      {column.render('Header')}
-                      {(data.length !== 0 || areFiltersSet) && (
-                        <FilterToggle
-                          column={column}
-                          ownerDocument={ownerDocument}
-                        />
-                      )}
-                      {data.length !== 0 && column.canSort && (
-                        <div className='iui-cell-end-icon'>
-                          {column.isSorted && column.isSortedDesc ? (
-                            <SvgSortDown
-                              className='iui-icon iui-sort'
-                              aria-hidden
-                            />
-                          ) : (
-                            <SvgSortUp
-                              className='iui-icon iui-sort'
-                              aria-hidden
-                            />
-                          )}
-                        </div>
-                      )}
-                      {isResizable &&
-                        column.isResizerVisible &&
-                        index !== headerGroup.headers.length - 1 && (
-                          <div
-                            {...column.getResizerProps()}
-                            className='iui-resizer'
-                          >
-                            <div className='iui-resizer-bar' />
+        <div className='iui-table-header-wrapper' ref={headerRef}>
+          <div className='iui-table-header'>
+            {headerGroups.slice(1).map((headerGroup: HeaderGroup<T>) => {
+              const headerGroupProps = headerGroup.getHeaderGroupProps({
+                className: 'iui-row',
+              });
+              return (
+                <div {...headerGroupProps} key={headerGroupProps.key}>
+                  {headerGroup.headers.map((column, index) => {
+                    const columnProps = column.getHeaderProps({
+                      ...column.getSortByToggleProps(),
+                      className: cx(
+                        'iui-cell',
+                        {
+                          'iui-actionable': column.canSort,
+                          'iui-sorted': column.isSorted,
+                          'iui-cell-sticky': !!column.sticky,
+                        },
+                        column.columnClassName,
+                      ),
+                      style: {
+                        ...getCellStyle(column, !!state.isTableResizing),
+                        ...getStickyStyle(column, visibleColumns),
+                        flexWrap: 'unset',
+                      },
+                    });
+                    return (
+                      <div
+                        {...columnProps}
+                        {...column.getDragAndDropProps()}
+                        key={columnProps.key}
+                        title={undefined}
+                        ref={(el) => {
+                          if (el) {
+                            columnRefs.current[column.id] = el;
+                            column.resizeWidth = el.getBoundingClientRect().width;
+                          }
+                        }}
+                      >
+                        {column.render('Header')}
+                        {(showFilterButton(column) ||
+                          showSortButton(column)) && (
+                          <div className='iui-table-header-actions-container'>
+                            {showFilterButton(column) && (
+                              <FilterToggle
+                                column={column}
+                                ownerDocument={ownerDocument}
+                              />
+                            )}
+                            {showSortButton(column) && (
+                              <div className='iui-cell-end-icon'>
+                                {column.isSorted && column.isSortedDesc ? (
+                                  <SvgSortDown
+                                    className='iui-icon iui-sort'
+                                    aria-hidden
+                                  />
+                                ) : (
+                                  <SvgSortUp
+                                    className='iui-icon iui-sort'
+                                    aria-hidden
+                                  />
+                                )}
+                              </div>
+                            )}
                           </div>
                         )}
-                      {enableColumnReordering && !column.disableReordering && (
-                        <div className='iui-reorder-bar' />
-                      )}
-                    </div>
-                  );
-                })}
-              </div>
-            );
-          })}
+                        {isResizable &&
+                          column.isResizerVisible &&
+                          index !== headerGroup.headers.length - 1 && (
+                            <div
+                              {...column.getResizerProps()}
+                              className='iui-resizer'
+                            >
+                              <div className='iui-resizer-bar' />
+                            </div>
+                          )}
+                        {enableColumnReordering &&
+                          !column.disableReordering && (
+                            <div className='iui-reorder-bar' />
+                          )}
+                        {column.sticky === 'left' &&
+                          state.sticky.isScrolledToRight && (
+                            <div className='iui-cell-shadow-right' />
+                          )}
+                        {column.sticky === 'right' &&
+                          state.sticky.isScrolledToLeft && (
+                            <div className='iui-cell-shadow-left' />
+                          )}
+                      </div>
+                    );
+                  })}
+                </div>
+              );
+            })}
+          </div>
         </div>
         <div
           {...getTableBodyProps({
@@ -760,6 +825,7 @@ export const Table = <
           onScroll={() => {
             if (headerRef.current && bodyRef.current) {
               headerRef.current.scrollLeft = bodyRef.current.scrollLeft;
+              updateStickyState();
             }
           }}
           tabIndex={-1}